RelatedArticleV2 to RelatedArticle
